refactor(website): clean up Product Stepper visuals structure and tokens
Extract icons, data constants, and types into co-located files to reduce
component file sizes. Consolidate shared design tokens, remove image
fallback logic, and strip section comments.
