+++ packages/twenty-website-new/src/sections/ProductStepper/types.ts
@@ -1,4 +1,3 @@
-import type { ImageType } from '@/design-system/components/Image';
 import type { MessageDescriptor } from '@lingui/core';
 import type { ComponentType, ReactNode } from 'react';
 
@@ -10,6 +9,5 @@ export type ProductStepperStepType = {
   body: MessageDescriptor;
   heading: ReactNode;
   icon: string;
-  image?: ImageType;
-  visual?: ComponentType<StepperVisualProps>;
+  visual: ComponentType<StepperVisualProps>;
 };
