+++ packages/twenty-website-new/src/sections/ProductStepper/visuals/stepper-visual-tokens.ts
@@ -1,24 +1,19 @@
 export const STEPPER_BG = '#ffffff';
-export const STEPPER_SURFACE = '#f9fafb';
-export const STEPPER_BORDER = '#e5e7eb';
-export const STEPPER_BORDER_LIGHT = '#f0f0f0';
-export const STEPPER_TEXT = '#1a1a2e';
+export const STEPPER_CARD_BG = '#fcfcfc';
+
+export const STEPPER_TEXT = '#333';
 export const STEPPER_TEXT_SECONDARY = '#6b7280';
 export const STEPPER_TEXT_TERTIARY = '#9ca3af';
+export const STEPPER_TEXT_MUTED = '#666';
+
+export const STEPPER_BORDER_MEDIUM = '#ebebeb';
+export const STEPPER_BORDER_STRONG = '#d6d6d6';
+export const STEPPER_BORDER_LIGHT = '#f1f1f1';
+export const STEPPER_BORDER_SUBTLE = 'rgba(0, 0, 0, 0.08)';
+
 export const STEPPER_FONT = "'Inter', sans-serif";
-export const STEPPER_RADIUS = '8px';
-export const STEPPER_RADIUS_SM = '4px';
-export const STEPPER_SHADOW = '0 1px 3px rgba(0, 0, 0, 0.06)';
+export const STEPPER_SHADOW_SM = '0 1px 6px rgba(0, 0, 0, 0.05)';
+export const STEPPER_TINT = 'rgba(0, 0, 0, 0.04)';
 
-export const STEPPER_ACCENT_BLUE = '#3b82f6';
-export const STEPPER_ACCENT_GREEN = '#22c55e';
-export const STEPPER_ACCENT_AMBER = '#f59e0b';
-export const STEPPER_ACCENT_PURPLE = '#8b5cf6';
-export const STEPPER_ACCENT_RED = '#ef4444';
-export const STEPPER_ACCENT_TEAL = '#14b8a6';
-
-export const STEPPER_ICON_BG_BLUE = 'rgba(59, 130, 246, 0.1)';
-export const STEPPER_ICON_BG_GREEN = 'rgba(34, 197, 94, 0.1)';
-export const STEPPER_ICON_BG_AMBER = 'rgba(245, 158, 11, 0.1)';
-export const STEPPER_ICON_BG_PURPLE = 'rgba(139, 92, 246, 0.1)';
-export const STEPPER_ICON_BG_RED = 'rgba(239, 68, 68, 0.1)';
+export const STEPPER_HEADER_BG = '#d9e2fc';
+export const STEPPER_HEADER_BORDER = '#c6d4f9';
